sheep: use long type for object cache size
This enables us to use more than 4 GB object cache on 64 bit machines.
This also works on i386 machines because long type is supported by
urcu on both x86_64 and i386.

+++ sheep/object_cache.c
@@ -48,7 +48,7 @@
 
 
 struct global_cache {
-	uint32_t cache_size;
+	unsigned long cache_size;
 	int in_reclaim;
 	struct cds_list_head cache_lru_list;
 };
@@ -515,7 +515,7 @@ static int do_reclaim_object(struct object_cache_entry *entry)
 		goto out;
 	}
 
-	dprintf("oid %"PRIx64" reclaimed successfully, cache_size: %"PRId32"\n",
+	dprintf("oid %"PRIx64" reclaimed successfully, cache_size: %ld\n",
 		oid, uatomic_read(&sys_cache.cache_size));
 	del_from_object_tree_and_list(entry, &oc->object_tree);
 out:
